backend/requirements/development.txt
# ============================================
# TruEditor - Geliştirme Bağımlılıkları
# ============================================

# Base requirements
-r base.txt

# Debug Tools
django-debug-toolbar>=4.2,<5.0
django-extensions>=3.2,<4.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
faker>=22.0,<30.0

# Code Quality
flake8>=7.0,<8.0
black>=24.0,<25.0
isort>=5.13,<6.0
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0

# Documentation
drf-spectacular>=0.27,<1.0

# Hot Reload
watchdog>=3.0,<4.0

# IPython for shell_plus
ipython>=8.20,<9.0
